final commit

+++ ft_isalnum.c
@@ -6,7 +6,7 @@
 /*   By: llopes-n <llopes-n@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/18 17:22:21 by llopes-n          #+#    #+#             */
-/*   Updated: 2021/08/30 19:48:11 by llopes-n         ###   ########.fr       */
+/*   Updated: 2021/09/09 01:28:35 by llopes-n         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 int	ft_isalnum(int c)
 {
-	return (ft_isalpha(c) || ft_isdigit(c));
+	return ((unsigned int)ft_isalpha(c) || (unsigned int)ft_isdigit(c));
 }
